ml/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
sentence-transformers==2.2.2
PyPDF2==3.0.1
pytube==15.0.0
youtube-transcript-api==0.6.1
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==1.5.3
accelerate==0.24.1
bitsandbytes==0.41.3
datasets==2.14.6